+++ source/solvers/navier_stokes_base.cc
@@ -15,7 +15,6 @@
 #include <solvers/postprocessing_cfd.h>
 #include <solvers/postprocessing_velocities.h>
 #include <solvers/postprocessors.h>
-#include <solvers/postprocessors_smoothing.h>
 
 #include <deal.II/distributed/fully_distributed_tria.h>
 #include <deal.II/distributed/grid_refinement.h>
@@ -2709,18 +2708,10 @@ NavierStokesBase<dim, VectorType, DofsType>::output_field_hook(
 
 template <int dim, typename VectorType, typename DofsType>
 void
-NavierStokesBase<dim, VectorType, DofsType>::write_output_results(
-  const VectorType &solution)
+NavierStokesBase<dim, VectorType, DofsType>::gather_output_results(
+  const VectorType                           &solution,
+  std::vector<OutputStruct<dim, VectorType>> &solution_output_structs)
 {
-  TimerOutput::Scope t(this->computing_timer, "Output VTU");
-
-  const std::string  folder        = simulation_control->get_output_path();
-  const std::string  solution_name = simulation_control->get_output_name();
-  const unsigned int iter          = simulation_control->get_step_number();
-  const double       time          = simulation_control->get_current_time();
-  const unsigned int subdivision = simulation_control->get_number_subdivision();
-  const unsigned int group_files = simulation_control->get_group_files();
-
   // Add the interpretation of the solution. The dim first components are the
   // velocity vectors and the following one is the pressure.
   std::vector<std::string> solution_names(dim, "velocity");
@@ -2731,20 +2722,17 @@ NavierStokesBase<dim, VectorType, DofsType>::write_output_results(
   data_component_interpretation.emplace_back(
     DataComponentInterpretation::component_is_scalar);
 
-  DataOut<dim> data_out;
+  OutputStructSolution<dim, VectorType> solution_fd_struct(
+    this->dof_handler, solution, solution_names, data_component_interpretation);
 
-  // Additional flag to enable the output of high-order elements
-  DataOutBase::VtkFlags flags;
-  if (this->velocity_fem_degree > 1)
-    flags.write_higher_order_cells = true;
-  data_out.set_flags(flags);
-
-  // Attach the solution data to data_out object
-  data_out.attach_dof_handler(this->dof_handler);
-  data_out.add_data_vector(solution,
-                           solution_names,
-                           DataOut<dim>::type_dof_data,
-                           data_component_interpretation);
+  // Add the interpretation of the solution to the output structs
+  // using in_place_type to save memory allocation and avoid unnecessary copies
+  solution_output_structs.emplace_back(
+    std::in_place_type<OutputStructSolution<dim, VectorType>>,
+    this->dof_handler,
+    solution,
+    solution_names,
+    data_component_interpretation);
 
   if (this->simulation_parameters.post_processing
         .calculate_average_velocities ||
@@ -2762,11 +2750,11 @@ NavierStokesBase<dim, VectorType, DofsType>::write_output_results(
           dim, DataComponentInterpretation::component_is_part_of_vector);
       average_data_component_interpretation.emplace_back(
         DataComponentInterpretation::component_is_scalar);
-
-      data_out.add_data_vector(
+      solution_output_structs.emplace_back(
+        std::in_place_type<OutputStructSolution<dim, VectorType>>,
+        this->dof_handler,
         this->average_velocities->get_average_velocities(),
         average_solution_names,
-        DataOut<dim>::type_dof_data,
         average_data_component_interpretation);
 
       // Add the interpretation of the reynolds stresses of solution.
@@ -2780,7 +2768,6 @@ NavierStokesBase<dim, VectorType, DofsType>::write_output_results(
           dim, DataComponentInterpretation::component_is_part_of_vector);
       reynolds_normal_stress_data_component_interpretation.emplace_back(
         DataComponentInterpretation::component_is_scalar);
-
 
       std::vector<std::string> reynolds_shear_stress_names = {
         "reynolds_shear_stress_uv"};
@@ -2801,33 +2788,42 @@ NavierStokesBase<dim, VectorType, DofsType>::write_output_results(
       reynolds_shear_stress_data_component_interpretation.emplace_back(
         DataComponentInterpretation::component_is_scalar);
 
-      data_out.add_data_vector(
+      solution_output_structs.emplace_back(
+        std::in_place_type<OutputStructSolution<dim, VectorType>>,
+        this->dof_handler,
         this->average_velocities->get_reynolds_normal_stresses(),
         reynolds_normal_stress_names,
-        DataOut<dim>::type_dof_data,
         reynolds_normal_stress_data_component_interpretation);
-
-      data_out.add_data_vector(
-        this->average_velocities->get_reynolds_shear_stresses(),
-        reynolds_shear_stress_names,
-        DataOut<dim>::type_dof_data,
-        reynolds_shear_stress_data_component_interpretation);
     }
-
-  Vector<float> subdomain(this->triangulation->n_active_cells());
-  for (unsigned int i = 0; i < subdomain.size(); ++i)
-    subdomain(i) = this->triangulation->locally_owned_subdomain();
-  data_out.add_data_vector(subdomain, "subdomain");
-
 
   // Create the post-processors to have derived information about the velocity
   // They are generated outside the if condition for smoothing to ensure
   // that the objects still exist when the write output of DataOut is called
   // Regular discontinuous postprocessors
-  QCriterionPostprocessor<dim> qcriterion;
-  DivergencePostprocessor<dim> divergence;
-  VorticityPostprocessor<dim>  vorticity;
-  data_out.add_data_vector(solution, vorticity);
+  // They are created as shared pointers to outlive the function
+  std::shared_ptr<QCriterionPostprocessor<dim>> qcriterion =
+    std::make_shared<QCriterionPostprocessor<dim>>();
+  solution_output_structs.emplace_back(
+    std::in_place_type<OutputStructPostprocessor<dim, VectorType>>,
+    this->dof_handler,
+    solution,
+    qcriterion);
+
+  std::shared_ptr<DivergencePostprocessor<dim>> divergence =
+    std::make_shared<DivergencePostprocessor<dim>>();
+  solution_output_structs.emplace_back(
+    std::in_place_type<OutputStructPostprocessor<dim, VectorType>>,
+    this->dof_handler,
+    solution,
+    divergence);
+
+  std::shared_ptr<VorticityPostprocessor<dim>> vorticity =
+    std::make_shared<VorticityPostprocessor<dim>>();
+  solution_output_structs.emplace_back(
+    std::in_place_type<OutputStructPostprocessor<dim, VectorType>>,
+    this->dof_handler,
+    solution,
+    vorticity);
 
   // Get physical properties models
   std::vector<std::shared_ptr<DensityModel>> density_models =
@@ -2840,23 +2836,26 @@ NavierStokesBase<dim, VectorType, DofsType>::write_output_results(
   const double n_fluids = this->simulation_parameters
                             .physical_properties_manager.get_number_of_fluids();
 
-  std::vector<DensityPostprocessor<dim>> density_postprocessors;
+  std::vector<std::shared_ptr<DensityPostprocessor<dim>>>
+    density_postprocessors;
   density_postprocessors.reserve(n_fluids);
-  std::vector<KinematicViscosityPostprocessor<dim>>
+  std::vector<std::shared_ptr<KinematicViscosityPostprocessor<dim>>>
     kinematic_viscosity_postprocessors;
   kinematic_viscosity_postprocessors.reserve(n_fluids);
-  std::vector<DynamicViscosityPostprocessor<dim>>
+  std::vector<std::shared_ptr<DynamicViscosityPostprocessor<dim>>>
     dynamic_viscosity_postprocessors;
   dynamic_viscosity_postprocessors.reserve(n_fluids);
 
   for (unsigned int f_id = 0; f_id < n_fluids; ++f_id)
     {
       density_postprocessors.emplace_back(
-        DensityPostprocessor<dim>(density_models[f_id], f_id));
+        std::make_shared<DensityPostprocessor<dim>>(density_models[f_id],
+                                                    f_id));
       kinematic_viscosity_postprocessors.emplace_back(
-        KinematicViscosityPostprocessor<dim>(rheological_models[f_id], f_id));
+        std::make_shared<KinematicViscosityPostprocessor<dim>>(
+          rheological_models[f_id], f_id));
       dynamic_viscosity_postprocessors.emplace_back(
-        DynamicViscosityPostprocessor<dim>(
+        std::make_shared<DynamicViscosityPostprocessor<dim>>(
           rheological_models[f_id],
           density_models[f_id]->get_density_ref(),
           f_id));
@@ -2866,90 +2865,43 @@ NavierStokesBase<dim, VectorType, DofsType>::write_output_results(
       if (!density_models[f_id]->is_constant_density_model() ||
           this->simulation_parameters.multiphysics.VOF ||
           this->simulation_parameters.multiphysics.cahn_hilliard)
-        data_out.add_data_vector(solution, density_postprocessors[f_id]);
+        solution_output_structs.emplace_back(
+          std::in_place_type<OutputStructPostprocessor<dim, VectorType>>,
+          this->dof_handler,
+          solution,
+          density_postprocessors[f_id]);
 
       // Only output the kinematic viscosity for non-newtonian rheology
       if (rheological_models[f_id]->is_non_newtonian_rheological_model())
         {
-          data_out.add_data_vector(solution,
-                                   kinematic_viscosity_postprocessors[f_id]);
+          solution_output_structs.emplace_back(
+            std::in_place_type<OutputStructPostprocessor<dim, VectorType>>,
+            this->dof_handler,
+            solution,
+            kinematic_viscosity_postprocessors[f_id]);
 
           // Only output the dynamic viscosity for multiphase flows
           if (this->simulation_parameters.multiphysics.VOF ||
               this->simulation_parameters.multiphysics.cahn_hilliard)
-            data_out.add_data_vector(solution,
-                                     dynamic_viscosity_postprocessors[f_id]);
+            solution_output_structs.emplace_back(
+              std::in_place_type<OutputStructPostprocessor<dim, VectorType>>,
+              this->dof_handler,
+              solution,
+              kinematic_viscosity_postprocessors[f_id]);
         }
     }
 
-  ShearRatePostprocessor<dim> shear_rate_processor;
   if (this->simulation_parameters.physical_properties_manager
         .is_non_newtonian())
-    data_out.add_data_vector(solution, shear_rate_processor);
-
-  // Trilinos vector for the smoothed output fields
-  QcriterionPostProcessorSmoothing<dim, VectorType> qcriterion_smoothing(
-    *this->triangulation,
-    this->simulation_parameters,
-    number_quadrature_points);
-
-  ContinuityPostProcessorSmoothing<dim, VectorType> continuity_smoothing(
-    *this->triangulation,
-    this->simulation_parameters,
-    number_quadrature_points);
-
-  // TODO: generalize this to VectorType
-  GlobalVectorType qcriterion_field;
-  GlobalVectorType continuity_field;
-
-  if (this->simulation_parameters.post_processing.smoothed_output_fields)
     {
-      // Qcriterion smoothing
-      {
-        qcriterion_field =
-          qcriterion_smoothing.calculate_smoothed_field(solution,
-                                                        this->dof_handler,
-                                                        this->get_mapping());
-
-        std::vector<DataComponentInterpretation::DataComponentInterpretation>
-          data_component_interpretation(
-            1, DataComponentInterpretation::component_is_scalar);
-
-        std::vector<std::string> qcriterion_name = {"qcriterion"};
-        const DoFHandler<dim>   &dof_handler_qcriterion =
-          qcriterion_smoothing.get_dof_handler();
-        data_out.add_data_vector(dof_handler_qcriterion,
-                                 qcriterion_field,
-                                 qcriterion_name,
-                                 data_component_interpretation);
-      }
-      // Continuity smoothing
-      {
-        continuity_field =
-          continuity_smoothing.calculate_smoothed_field(solution,
-                                                        this->dof_handler,
-                                                        this->get_mapping());
-
-        std::vector<DataComponentInterpretation::DataComponentInterpretation>
-          data_component_interpretation(
-            1, DataComponentInterpretation::component_is_scalar);
-
-        std::vector<std::string> continuity_name = {"velocity_divergence"};
-        const DoFHandler<dim>   &dof_handler_qcriterion =
-          continuity_smoothing.get_dof_handler();
-        data_out.add_data_vector(dof_handler_qcriterion,
-                                 continuity_field,
-                                 continuity_name,
-                                 data_component_interpretation);
-      }
+      auto shear_rate_processor =
+        std::make_shared<ShearRatePostprocessor<dim>>();
+      solution_output_structs.emplace_back(
+        std::in_place_type<OutputStructPostprocessor<dim, VectorType>>,
+        this->dof_handler,
+        solution,
+        shear_rate_processor);
     }
-  else
-    {
-      // Use the non-smoothed version of the post-processors
-      data_out.add_data_vector(solution, divergence);
-      data_out.add_data_vector(solution, qcriterion);
-    }
-
 
   SRFPostprocessor<dim> srf(simulation_parameters.velocity_sources.omega_x,
                             simulation_parameters.velocity_sources.omega_y,
@@ -2957,7 +2909,75 @@ NavierStokesBase<dim, VectorType, DofsType>::write_output_results(
 
   if (simulation_parameters.velocity_sources.rotating_frame_type ==
       Parameters::VelocitySource::RotatingFrameType::srf)
-    data_out.add_data_vector(solution, srf);
+    solution_output_structs.emplace_back(
+      std::in_place_type<OutputStructPostprocessor<dim, VectorType>>,
+      this->dof_handler,
+      solution,
+      std::make_shared<SRFPostprocessor<dim>>(srf));
+}
+
+template <int dim, typename VectorType, typename DofsType>
+void
+NavierStokesBase<dim, VectorType, DofsType>::write_output_results(
+  const VectorType &solution)
+{
+  TimerOutput::Scope t(this->computing_timer, "Output results");
+
+  const std::string  folder        = simulation_control->get_output_path();
+  const std::string  solution_name = simulation_control->get_output_name();
+  const unsigned int iter          = simulation_control->get_step_number();
+  const double       time          = simulation_control->get_current_time();
+  const unsigned int subdivision = simulation_control->get_number_subdivision();
+  const unsigned int group_files = simulation_control->get_group_files();
+
+  // Vector containing structs with vtu output information
+  std::vector<OutputStruct<dim, VectorType>> solution_output_structs;
+
+  // Gather all results in vector container
+  gather_output_results(this->present_solution, solution_output_structs);
+
+  // Create data output object
+  DataOut<dim> data_out;
+
+  // Additional flag to enable the output of high-order elements
+  DataOutBase::VtkFlags flags;
+  if (this->velocity_fem_degree > 1)
+    flags.write_higher_order_cells = true;
+  data_out.set_flags(flags);
+
+  // Fill data out object with solutions in structs
+  for (const auto &solution_output_struct : solution_output_structs)
+    {
+      if (auto solution_struct =
+            std::get_if<OutputStructSolution<dim, VectorType>>(
+              &solution_output_struct))
+        {
+          // auto solution_output = *solution_struct;
+          data_out.add_data_vector(
+            solution_struct->dof_handler,
+            solution_struct->solution,
+            solution_struct->solution_names,
+            solution_struct->data_component_interpretation);
+        }
+      else if (auto postprocessor_struct =
+                 std::get_if<OutputStructPostprocessor<dim, VectorType>>(
+                   &solution_output_struct))
+        {
+          data_out.add_data_vector(postprocessor_struct->dof_handler,
+                                   postprocessor_struct->solution,
+                                   *postprocessor_struct->data_postprocessor);
+        }
+    }
+
+  // Since the subdomain is a cell data, we process it separately to avoid
+  // storing DataOut<dim>::type_dof_data for every DoF field in the solution
+  // output struct container.
+  Vector<float> subdomain(this->triangulation->n_active_cells());
+  for (unsigned int i = 0; i < subdomain.size(); ++i)
+    subdomain(i) = this->triangulation->locally_owned_subdomain();
+  data_out.add_data_vector(subdomain,
+                           "subdomain",
+                           DataOut<dim>::type_cell_data);
 
   output_field_hook(data_out);
 
@@ -2976,6 +2996,7 @@ NavierStokesBase<dim, VectorType, DofsType>::write_output_results(
                          iter,
                          group_files,
                          this->mpi_communicator);
+
 
   if (simulation_control->get_output_boundaries() &&
       simulation_control->get_step_number() == 0)